requirements.txt
tensorflow
numpy
python-dotenv
opencv-python